+++ front/src/helpers/validatePassword.ts
@@ -1,0 +1,5 @@
+export function validatePassword(password: string): boolean {
+  return /^(?=[a-zA-Z0-9]{8,}$)(?=.*?[a-z])(?=.*?[A-Z])(?=.*?\d).*/.test(
+    password
+  )
+}