+++ front/src/hooks/useRegister.tsx
@@ -1,25 +1,23 @@
 import { useMutation } from '@tanstack/react-query'
-import { useState } from 'react'
 import { registerUserFetcher } from '../helpers/fetchers'
+import { type TRegisterForm } from '../types'
 
 export const useRegister = (handleRegisterModal: () => void) => {
-  const [responseError, setResponseError] = useState('')
-  const registerUser = useMutation(registerUserFetcher, {
-    onMutate: () =>
-      new Promise((resolve) => {
-        setTimeout(resolve, 500)
-      }),
-    onSuccess: () => {
-      setTimeout(handleRegisterModal, 2000)
-    },
-    onError: (error: Error) => {
-      setResponseError(error.message)
-    },
-  })
+  const registerUser = useMutation<null, Error, TRegisterForm>(
+    registerUserFetcher,
+    {
+      onMutate: () =>
+        new Promise((resolve) => {
+          setTimeout(resolve, 500)
+        }),
+      onSuccess: () => {
+        setTimeout(handleRegisterModal, 2000)
+      },
+    }
+  )
   return {
     registerUser,
-    responseError,
-    setResponseError,
+    error: registerUser.error,
     isLoading: registerUser.isLoading,
     isSuccess: registerUser.isSuccess,
   }