+++ sso/src/__tests__/globalSetup.ts
@@ -41,34 +41,34 @@ export const testUserData = {
 }
 
 export async function setup() {
-  await client.query('DROP TABLE IF EXISTS users CASCADE')
+  await client.query('DROP TABLE IF EXISTS "user" CASCADE')
   await client.query(`
-    CREATE OR REPLACE FUNCTION TRIGGER_SET_TIMESTAMP() RETURNS TRIGGER AS $$ BEGIN NEW.updated_at := NOW();
-    RETURN NEW;
-    END;
-    $$LANGUAGE plpgsql;
-    
-    CREATE TABLE IF NOT EXISTS users (
-        id VARCHAR (50) PRIMARY KEY,
-        dni VARCHAR(25) UNIQUE NOT NULL,
-        email VARCHAR (255) UNIQUE NOT NULL,
-        password TEXT NOT NULL,
-        user_meta JSON NOT NULL,
-        created_at TIMESTAMPTZ DEFAULT NOW (),
-        updated_at TIMESTAMPTZ
-    );
-    
-    CREATE TRIGGER set_timestamp BEFORE
-    UPDATE ON users FOR EACH ROW EXECUTE FUNCTION trigger_set_timestamp();
+  CREATE OR REPLACE FUNCTION TRIGGER_SET_TIMESTAMP() RETURNS TRIGGER AS $$ BEGIN NEW.updated_at := NOW();
+RETURN NEW;
+END;
+$$LANGUAGE plpgsql;
+
+CREATE TABLE IF NOT EXISTS "user" (
+    id TEXT PRIMARY KEY,
+    dni VARCHAR(25) UNIQUE NOT NULL,
+    email VARCHAR (255) UNIQUE NOT NULL,
+    password TEXT NOT NULL,
+    user_meta JSONB NOT NULL DEFAULT '{}',
+    created_at TIMESTAMPTZ DEFAULT NOW (),
+    updated_at TIMESTAMPTZ
+);
+
+CREATE TRIGGER set_timestamp BEFORE
+UPDATE ON "user" FOR EACH ROW EXECUTE FUNCTION trigger_set_timestamp();
 `)
 
   // Cleanup database
-  await client.query('DELETE FROM users')
+  await client.query('DELETE FROM "user"')
 
   // Create required test data
-  Object.values(testUserData).forEach(async (user) => {
+  const userCreationPromises = Object.values(testUserData).map(async (user) => {
     const query = `
-      INSERT INTO users (id, dni, email, password, user_meta)
+      INSERT INTO "user" (id, dni, email, password, user_meta)
       VALUES ($1, $2, $3, $4, $5)
       ON CONFLICT (dni) DO NOTHING;`
 
@@ -80,11 +80,13 @@ export async function setup() {
       user.user_meta,
     ])
   })
+
+  await Promise.all(userCreationPromises)
 }
 
 export async function teardown() {
   // Cleanup database
-  await client.query('DELETE FROM users')
+  await client.query('DELETE FROM "user"')
   await client.end()
   server.close()
 }