+++ front/CHANGELOG.md
@@ -3,11 +3,26 @@
 All notable changes to this project will be documented in this file.
 
 
-## [0.5.0] - 2023-11-22
+## [0.5.0] - 2023-11-23
 
 ### Added
 
 - New resource search bar UI implementation
+
+## [0.4.4] - 2023-11-23
+
+### Fixed
+
+- Adjusted the general styles of the AdminAccount table.
+- Updated translations accordingly.
+- Relocated AccountAdmin.test from the "pages" directory to "/components/organisms".
+
+
+## [0.4.3] - 2023-11-23
+
+### Fixed
+
+- Fix register error message.
 
 ## [0.4.2] - 2023-11-22